refactor: remove conditional enablement of our action
It's just confusing in my opinion. Previously we were using the approach
shown in the Chrome extension docs to conditionally enable our action
(i.e. the button on the nav bar) when the tab was a Reddit tab.

That is the effect we want, as the popup only works on Reddit, but the
way it works in the browser (at least in Brave) is that the action
starts out greyed out, then once you open a Reddit tab, it becomes
coloured, i.e. active. Thereafter it remains active, even on non-reddit
tabs. When clicked, the action shows the right-click context menu.

It's better for us to just open the popup so we can show a message
informing the user that they need to go to a Reddit tab.