refactor: use messages to fetch avatar data
We now use chrome.runtime.onMessage / chrome.tabs.sendMessage to fetch
avatar data from our code running in a Reddit tab. Prevoiusly the
injected script was set up to return a Promise with the data. That works
fine, but required some bodging of the Webpack build, is less easy to
test, and is a less general solution.