fix: prevent non-NFT avatars being rendered as NFT Cards
Non-NFT avatars were being rendered with the NFT Card style if the NFT
Card style option was restored from storage and the current Avatar was
not an NFT Avatar. The style now defaults to the Standard style if an
incompatible or unimplemented style is requested.

This fixes https://github.com/h4l/headgear/issues/5