feat: use a vector mask to clip NFT Card bg image
With this change, the pixelated aliasing of the NFT Card background
image's corners is no longer apparent when scaled greater than the
background image's 1:1 size.

We now use a rounded rectangle vector mask to clip the NFT Card's
background image, instead of relying on the PNG's alpha channel. The
comment in the SVG template explains the approach, as it's more than
just clipping the <image/>.