feat: async-generate output PNGs
A new state type - OutputImageState - represents the result of
generating the Avatar image to be downloaded/copied. It can represent
either PNGs or SVGs depending on the image settings chosen. As with the
other state objects, we use a Signal to represent the current state,
and automatically update it as dependencies change (image settings,
avatar image style, etc).