feat: add stand-alone text-to-path service
This service can generate SVG paths from strings of text, with custom
fonts. We'll use it to remove the need to embed reference/embed web
fonts in our generated SVG files.

It's implemented as a separate service in order to maintain a clear
boundary between this portion of the codebase and the rest. The reason
for that is that we use opentype.js for text to path rendering, and this
library is massive (—0.5MB of js) and I want the code to remain
auditable. Bundling so much code into the main part of the app goes
against that goal.