feat: remove need for web fonts in SVG
We now generate SVG <path>s to represent text that previously used web
fonts. This removes all external dependencies, so SVG docs are entirely
stand-alone. It also allows them to be rendered correctly by programs
not capable of rendering web fonts in SVG (e.g. Inkscape).

A further benefit is that Firefox can render these SVG docs to a canvas,
without the need to embed the fonts as data: URIs (FF doesn't allow
external references in SVG files rendered to a canvas).

We keep an invisible <text> element at the same position as the visible
<path> text, so the text is still present in the SVG file for metadata
& copying to clipboard purposes.