refactor: get rid of useValueHistory()
Don't need it, instead we now directly determine whether a transition
is needed via the wasClosed state. This also avoids a slight flash when
re-opening Headgear's popup when a modal was active, due to a frame of
delay where the modal background renders without the modal itself
rendering, due to a delay in propagating the values through
useValueHistory()'s state.