fix: prevent shadow clipping on NFT Card text
We had some clipping of the text shadow at the top and bottom after
after changing the text to paths from real text. The previous filter and
mask applied to the text was sized relative to the target object's
bounding box, which must be tighter for <path> compared to <text>, which
resulted in clipping. We now size the filter & mask relative to the
containing (sub) svg element, which provides a generous but reasonable
bounding box.

The "1" in the series size does now have an over-large filter area, as
it uses the same region as the main name text. Probably not worth
optimising to reduce overdraw though... I think it'd require re-defining
the shadow filter.