chore: update Avatar Data GQL query ID
Not sure how frequently this changes, but it's nice of Reddit to keep
the old one working. 🥰 I guess it was last changed when they introduced
the feature allowing specific backgrounds to be chosen, because the
current response seems to explicitly specify the background used on a
user avatar. So we can probably get rid of the previous hack that parsed
/ base64-decoded the pre-composed avatar image URL to extract the NFT
ID. For now the old method is still compatible though.