fix(analytics): ensure analytics is enabled for opted-in users
Previously, analytics events were not actually active if a user opted-in
after having started from an opted-out state. This was because the logic
assumed that calling init() on a PostHog instance would also
enable/opt-in analytics, but PostHog requires an explicit opt-in after
having been opted-out.

We now check the PostHog opt-in/out state when opting in, and opt in
explicitly if required.

This fixes https://github.com/h4l/headgear/issues/29