feat: fetch user auth token from cookies
Reddit changed the content of the /coins and /premium pages we used to
use to access user auth tokens. It seems like the token is now not
available from page data, so we instead access it from the `token_v2`
cookie.

This does mean that we need to increase the permission scope to include
cookies. We could use optional host permissions and request access on
first opening, but doing that would take a decent amount more dev effort
than just changing permissions. The user base is pretty small, so the
impact of users needing to re-approve the extension before using the
updated version seems like an OK trade off to ship this out more
quickly, with less dev time required.