feat: invalidate cached avatar if user changes
We now invalidate the cached avatar data when a user logs out. We do
this by incorporating a subset of the auth token into the value that is
hashed to form the cache version ID. This does mean that the cache will
invalidate when the token rotates, but that shouldn't happen more than
once a day.